gateway/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-jose==3.5.0
httpx==0.28.1
pydantic-settings==2.10.1